feat(user-test): implement result service

+++ api-gateway/src/main/java/com/nitb/apigateway/dto/UserTest/response/ResultSummaryResponseDto.java
@@ -21,5 +21,5 @@ public class ResultSummaryResponseDto {
 
     private int score;
 
-    private int secondsSpent;
+    private long secondsSpent;
 }
